chore: separate data
